+++ src/components/UserPage/PetsData/PetsData.jsx
@@ -6,7 +6,13 @@ import { ModalAddsPet } from '../ModalAddPets/ModalAddPets';
 
 import { ReactComponent as PlusIcon } from '../../../assets/icons/plusIcon.svg';
 
-import { Container, AddPetBtn, Title, Text } from './PetsData.styled';
+import {
+  Container,
+  AddPetBtn,
+  Title,
+  Text,
+  AddPetBoxWrapper,
+} from './PetsData.styled';
 
 export const PetsData = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -19,12 +25,12 @@ export const PetsData = () => {
     <Container>
       <Box display="flex" justifyContent="space-between" alignItems="center">
         <Title>My pets:</Title>
-        <Box display="flex" alignItems="center">
+        <AddPetBoxWrapper>
           <Text>Add pet</Text>
           <AddPetBtn type="button" onClick={() => setIsModalOpen(true)}>
             <PlusIcon />
           </AddPetBtn>
-        </Box>
+        </AddPetBoxWrapper>
       </Box>
       {isModalOpen && (
         <Modal onClick={closeModal}>
